First pass at serialize reading.
